+++ kubejs/server_scripts/Mods/Greate/Press.js
@@ -1,8 +1,4 @@
 ServerEvents.recipes(event => {
-    event.recipes.greate.pressing('kubejs:unf_turtle_helmet', '5x minecraft:scute')
-		.circuitNumber(1)
-        .recipeTier(1)
-
 	event.recipes.greate.pressing('kubejs:ruthenium_trinium_americium_neutronate_plate', 'gtceu:ruthenium_trinium_americium_neutronate_ingot')
 		.recipeTier(1)
 		.circuitNumber(1)