+++ kubejs/server_scripts/Mods/Greg/Bender.js
@@ -1,0 +1,7 @@
+ServerEvents.recipes(event => {
+    event.recipes.gtceu.bender('unf_turtle_helmet')
+        .itemInputs('5x minecraft:scute')
+        .itemOutputs('kubejs:unf_turtle_helmet')
+		.circuit(1)
+        .EUt(30)
+})